requirements.txt
streamlit
pandas
requests
beautifulsoup4
pytest
python-dotenv
openai
google-generativeai
anthropic
cloudscraper
